Feature/transactions hook

+++ server/migrations/20230531005729-create-account.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       totalValue: {
-        type: Sequelize.BIGINT
+        type: Sequelize.DOUBLE
       },
       userId: {
         allowNull: false,
